# api/openapi.yaml
openapi: 3.0.3
info:
  title: CompAInion API
  version: 0.1.0
  description: |
    REST backend for CompAInion: a Hermes session acting as "companion"
    dispatches plans/tasks to detached Hermes workers and serves the user one
    interruption at a time.

    Conventions:
    - Times are RFC3339 UTC strings.
    - JSON-valued columns (`handle`, `scope`, `acceptance_checks`, `evidence`,
      `payload`, `outcome`) are typed as their natural JSON shape here.
    - Optimistic concurrency: plans/steps/tasks carry `version`; PATCH takes
      `If-Match: "<version>"` (missing -> 428, stale -> 412). Responses carry an
      `ETag: "<version>"` header.
    - Every creating `POST` accepts `Idempotency-Key`; replay returns the cached
      response with `Idempotent-Replayed: true`, same key + different body -> 422.
    - Lists are cursor paginated: `?cursor=<last id>&limit=<n>` (default 50, max
      200), envelope `{items, next_cursor}`.
    - Errors are RFC 7807 `application/problem+json` (400/404/409/412/422/428).
    - Step/task `status` is the stored status; `ready` is a derived boolean
      (pending + all deps done). A `ready` value is exposed ONLY via this flag.
servers:
  - url: http://127.0.0.1:7777
...
paths:
...
  /plans/{id}/steps:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: listPlanSteps
      tags: [steps, plans]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated steps (blocked ones show `ready:false`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StepPage' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: createStep
      tags: [steps]
      description: |
        The plan must be `draft` or `running`. `kind=goal` is rejected on
        create (only the plan creates it). `added_by` is required. `scope`
        must be narrower than the plan scope. Deps must be same-plan and must
        not introduce a cycle. A new step with no dependents automatically
        becomes a dependency of the `goal` step (the DAG ends in goal).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StepInput' }
      responses:
        '201':
          description: Created (with `ETag`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Step' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
  /steps/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getStep
      tags: [steps]
      responses:
        '200':
          description: The step (with `ETag`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Step' }
        '404': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateStep
      tags: [steps]
      description: Only while stored status is `pending` (else 409). The goal
        step needs `approval_ref` in the body (else 428). Requires `If-Match`.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StepInput' }
      responses:
        '200':
          description: Updated (with `ETag`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Step' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: deleteStep
      tags: [steps]
      description: Only `pending` steps that are not the goal step. The goal
        step can never be deleted (409 even with `approval_ref`).
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /steps/{id}/claim:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: claimStep
      tags: [steps]
      description: 'pending + ready -> in_progress; sets `assignee_agent_id`,'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_id]
              properties:
                agent_id: { type: string }
      responses:
        '200':
          description: Claimed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Step' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /steps/{id}/finish:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: finishStep
      tags: [steps]
      description: |
        Transitions `in_progress -> done|failed|blocked|interrupted`. `outcome`
        is required and must carry `result`; `done` requires a non-empty
        `outcome.evidence` (else 422). Completing the `goal` step sets the plan
        `done` - but only if all other steps are `done`/`cancelled` (else 409).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, outcome]
              properties:
                status:
                  type: string
                  enum: [done, failed, blocked, interrupted]
                outcome: { $ref: '#/components/schemas/Outcome' }
            example:
              status: done
              outcome:
                result: All unit tests pass; lint clean.
                evidence: ['go test ./... -> ok', 'lint output empty']
      responses:
        '200':
          description: Finished
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Step' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /steps/{id}/deps:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      operationId: addStepDep
      tags: [steps]
      description: Same editing rules as step update; `depends_on_id` must be a
        step in the same plan and adding it must not create a cycle.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DependsOn' }
      responses:
        '204': { description: Added }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /steps/{id}/deps/{dep_id}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - name: dep_id
        in: path
        required: true
        schema: { type: string }
    delete:
      operationId: removeStepDep
      tags: [steps]
      responses:
        '204': { description: Removed }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      description: |
        Append-only event log ordered by ULID id. `since` is an exclusive
        cursor (event id); `limit` default 100, max 500; order ascending;
        `next_cursor` is the last id. `wait=<seconds>` long-polls for new
        events (returns an empty page on timeout).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - { name: since, in: query, schema: { type: string } }
        - name: wait
          in: query
          description: Long-poll up to this many seconds.
          schema: { type: integer, minimum: 0, maximum: 60 }
        - { name: agent_id, in: query, schema: { type: string } }
        - { name: plan_id, in: query, schema: { type: string } }
        - { name: step_id, in: query, schema: { type: string } }
        - { name: task_id, in: query, schema: { type: string } }
        - { name: type, in: query, description: Repeatable., schema: { type: string } }
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventPage' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: appendEvent
      tags: [events]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '201':
          description: Appended
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        default: { $ref: '#/components/responses/Problem' }
  /stream:
    get:
      operationId: streamEvents
      tags: [events]
      description: |
        Server-sent events (signal only). Each message is
        `id: <event id>\nevent: <type>\ndata: {"type":..,"id":..}\n\n`.
        Honours `Last-Event-ID` (or the `since` query) by first replaying log
        events with id greater than it. Heartbeat comment `: ping` every 15s.
      parameters:
        - name: types
          in: query
          description: Comma-separated event types to filter on.
          schema: { type: string }
        - { name: workstream_id, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string } }
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  id: 01JABCDEVENT00000000000
                  event: needs_input
                  data: {"type":"needs_input","id":"01JABCDEVENT00000000000"}

                  : ping
        default: { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor (the last id of the previous page).
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: Page size (default 50, max 200 for resources; 100/500 for events).
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IfMatch:
      name: If-Match
      in: header
      description: 'Expected `version` as an ETag, e.g. "3". Required on PATCH'
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Client-supplied key caching the response of this creating
        POST. Same key + different body -> 422.
      schema: { type: string }
  responses:
    Problem:
      description: RFC 7807 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details. `type` maps the domain error kind
        (not_found, conflict, invalid, unprocessable, precondition_failed,
        precondition_required).
      required: [type, title, status]
      properties:
        type: { type: string, example: not_found }
        title: { type: string, example: Resource not found }
        status: { type: integer, example: 404 }
        detail: { type: string, example: plan 01J... not found }
        instance: { type: string }
...
    WorkStatus:
      type: string
      description: Stored step/task status; `ready` is only ever the derived
        boolean flag, never a status value.
      enum: [pending, in_progress, done, failed, blocked, interrupted, cancelled]
    StepKind:
      type: string
      enum: [task, checkpoint, goal]
    AddedBy:
      type: string
      enum: [planner, orchestrator, user]
    Step:
      type: object
      required: [id, plan_id, kind, title, description, acceptance_criteria,
        scope, status, attempt, added_by, version, created_at, updated_at]
      properties:
        id: { type: string }
        plan_id: { type: string }
        kind: { $ref: '#/components/schemas/StepKind' }
        title: { type: string }
        description: { type: string }
        acceptance_criteria: { type: string }
        scope: { type: object, additionalProperties: true }
        status: { $ref: '#/components/schemas/WorkStatus' }
        ready:
          type: boolean
          description: Derived on read (pending + all deps done); never stored.
        assignee_agent_id: { type: string, nullable: true }
        outcome:
          allOf:
            - $ref: '#/components/schemas/Outcome'
          nullable: true
        attempt: { type: integer }
        added_by: { $ref: '#/components/schemas/AddedBy' }
        suggested_executor: { type: string }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deps:
          type: array
          description: Ids of steps this step depends on.
          items: { type: string }
    StepInput:
      type: object
      required: [kind, title, added_by]
      properties:
        kind: { $ref: '#/components/schemas/StepKind' }
        title: { type: string }
        description: { type: string }
        acceptance_criteria: { type: string }
        scope: { type: object, additionalProperties: true }
        added_by: { $ref: '#/components/schemas/AddedBy' }
        suggested_executor: { type: string }
        deps:
          type: array
          description: Ids of same-plan steps this step depends on (cycle-checked).
          items: { type: string }
        approval_ref:
          type: string
          description: Required when editing the goal step.
        depends_on_id:
          type: string
          description: For POST /steps/{id}/deps.
    DependsOn:
      type: object
      required: [depends_on_id]
      properties:
        depends_on_id: { type: string }
    Outcome:
      type: object
      required: [result]
      properties:
        result: { type: string }
        evidence:
          type: array
          description: Required non-empty when finishing with `done`.
          items: { type: string }
...
    EventType:
      type: string
      enum: [started, progress, needs_input, finished, steer, error, note]
    Event:
      type: object
      required: [id, type, payload, at]
      properties:
        id:
          type: string
          description: ULID; also the list cursor.
        agent_id: { type: string }
        plan_id: { type: string }
        step_id: { type: string }
        task_id: { type: string }
        type: { $ref: '#/components/schemas/EventType' }
        payload:
          type: object
          additionalProperties: true
          description: JSON object. Synthetic companion signals ride
            `type=note` with `payload.kind` set to
            `interruption_created|agent_waiting|agent_lost|interruption_answered`.
        at: { type: string, format: date-time }
    EventInput:
      type: object
      required: [type]
      properties:
        agent_id: { type: string }
        plan_id: { type: string }
        step_id: { type: string }
        task_id: { type: string }
        type: { $ref: '#/components/schemas/EventType' }
        payload: { type: object, additionalProperties: true }
...
    Page:
      type: object
      description: Standard list envelope; `next_cursor` is null on the last page.
      required: [items]
      properties:
        next_cursor: { type: string, nullable: true }
...
    StepPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Step' }
...
    EventPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Event' }
...